refactor into multiple files in a new dir for clarity
